+++ src/api/v2/events.rs
@@ -168,7 +168,7 @@ mod tests {
   use futures::StreamExt;
   use futures::TryStreamExt;
 
-  use test_env_log::test;
+  use test_log::test;
 
   use url::Url;