+++ src/api/v2/position.rs
@@ -146,7 +146,7 @@ mod tests {
 
   use serde_json::from_str as from_json;
 
-  use test_env_log::test;
+  use test_log::test;
 
   use crate::api_info::ApiInfo;
   use crate::Client;