+++ src/api/v2/positions.rs
@@ -31,7 +31,7 @@ Endpoint! {
 mod tests {
   use super::*;
 
-  use test_env_log::test;
+  use test_log::test;
 
   use crate::api_info::ApiInfo;
   use crate::Client;