+++ src/data/v2/mod.rs
@@ -5,7 +5,3 @@
 pub mod bars;
 /// Functionality for retrieval of most recent quotes.
 pub mod last_quote;
-
-// TODO: Remove this alias with the next compatibility breaking release.
-#[deprecated(note = "renamed to 'bars'; use that instead")]
-pub use bars as stocks;