+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Unreleased
 - Removed support for streaming account updates
 - Switched from using `test-env-log` to `test-log`
 - Bumped minimum supported Rust version to `1.51`
-- Bumped `websocket-util` dependency to `0.9`
+- Bumped `websocket-util` dependency to `0.10`
 - Bumped `tokio-tungstenite` dependency to `0.16`