+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Unreleased
   `api::v2::order::PatchError` to `NotPermitted`
 - Removed support for streaming account updates
 - Switched from using `test-env-log` to `test-log`
-- Bumped minimum supported Rust version to `1.51`
+- Bumped minimum supported Rust version to `1.56`
 - Bumped `websocket-util` dependency to `0.10`
 - Bumped `tokio-tungstenite` dependency to `0.16`